Move source icons into subfolder
