+++ x-pack/plugins/enterprise_search/public/applications/workplace_search/components/shared/assets/source_icons/index.ts
