+++ x-pack/plugins/enterprise_search/public/applications/workplace_search/components/shared/source_icon/source_icon.tsx
@@ -10,7 +10,7 @@ import React from 'react';
 // eslint-disable-next-line no-restricted-imports
 import _camelCase from 'lodash/camelCase';
 
-import { images } from '../assets';
+import { images } from '../assets/source_icons';
 
 interface SourceIconProps {
   serviceType: string;
